+++ lib/core/cardano-wallet-core.cabal
@@ -68,7 +68,6 @@ library
   hs-source-dirs:
       src
   exposed-modules:
-      Cardano.Environment
       Cardano.Wallet
       Cardano.Wallet.Api
       Cardano.Wallet.Api.Server
@@ -78,6 +77,7 @@ library
       Cardano.Wallet.DB.Sqlite
       Cardano.Wallet.DB.Sqlite.TH
       Cardano.Wallet.DB.Sqlite.Types
+      Cardano.Wallet.Environment
       Cardano.Wallet.Network
       Cardano.Wallet.Primitive.AddressDerivation
       Cardano.Wallet.Primitive.AddressDiscovery
@@ -148,7 +148,7 @@ test-suite unit
       Cardano.Wallet.DB.MVarSpec
       Cardano.Wallet.DB.SqliteSpec
       Cardano.Wallet.DBSpec
-      Cardano.EnvironmentSpec
+      Cardano.Wallet.EnvironmentSpec
       Cardano.Wallet.NetworkSpec
       Cardano.Wallet.Primitive.AddressDerivationSpec
       Cardano.Wallet.Primitive.AddressDiscoverySpec
